example/requirements.txt
django>=1.11.0,<=1.11.99
django-prices>=0.7,<0.8
django-mptt==0.9.0  # for builtin categoies only
prices>=0.5,<0.6
satchless>=1.1.2,<1.2
jsonfield==2.0.2
